agent_missions_api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.25.0
opentelemetry-api>=1.36.0,<1.39.0
opentelemetry-sdk>=1.36.0,<1.39.0
opentelemetry-exporter-otlp-proto-grpc>=1.36.0,<1.39.0
opentelemetry-exporter-otlp-proto-http>=1.36.0,<1.39.0
opentelemetry-exporter-gcp-trace>=1.6.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-redis
mcp[sse,fastapi]>=1.0.0
google-adk==1.28.1
google-genai>=1.0.0
litellm>=1.0.0
prometheus-fastapi-instrumentator>=6.1.0
redis>=5.0.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
google-cloud-logging>=3.8.0
python-json-logger>=2.0.7
tenacity
